// src/common/constants/response-messages.enum.ts
export const RESPONSE_EXPLAINATION = {
  LIST:
    'First element is the array of result, second element is the total count',
  LIST_SPECIAL:
    'First element is result of the list, second is the total count, last is unread count',
};